chore: remove the legacy ios code to avoid confusion and standardize on catalogs
